Adapters: share common imp utils
