+++ modules/adxcgBidAdapter.js
@@ -3,7 +3,6 @@ import { ortbConverter } from '../libraries/ortbConverter/converter.js';
 import { registerBidder } from '../src/adapters/bidderFactory.js';
 import { BANNER, NATIVE, VIDEO } from '../src/mediaTypes.js';
 import {
-  isArray,
   replaceAuctionPrice,
   triggerPixel,
   logMessage,
@@ -11,6 +10,7 @@ import {
   getBidIdParameter
 } from '../src/utils.js';
 import { config } from '../src/config.js';
+import { applyCommonImpParams } from '../libraries/impUtils.js';
 
 const BIDDER_CODE = 'adxcg';
 const SECURE_BID_URL = 'https://pbc.adxcg.net/rtb/ortb/pbc?adExchangeId=1';
@@ -101,26 +101,7 @@ const converter = ortbConverter({
     const imp = buildImp(bidRequest, context);
     // tagid
     imp.tagid = bidRequest.params.adzoneid.toString();
-    // unknown params
-    const unknownParams = slotUnknownParams(bidRequest);
-    if (imp.ext || unknownParams) {
-      imp.ext = Object.assign({}, imp.ext, unknownParams);
-    }
-    // battr
-    if (bidRequest.params.battr) {
-      ['banner', 'video', 'audio', 'native'].forEach(k => {
-        if (imp[k]) {
-          imp[k].battr = bidRequest.params.battr;
-        }
-      });
-    }
-    // deals
-    if (bidRequest.params.deals && isArray(bidRequest.params.deals)) {
-      imp.pmp = {
-        private_auction: 0,
-        deals: bidRequest.params.deals
-      };
-    }
+    applyCommonImpParams(imp, bidRequest, KNOWN_PARAMS);
 
     imp.secure = bidRequest.ortb2Imp?.secure ?? 1;
 
@@ -147,20 +128,5 @@ const converter = ortbConverter({
     return bidResponse;
   },
 });
-
-/**
- * Unknown params are captured and sent on ext
- */
-function slotUnknownParams(slot) {
-  const ext = {};
-  const knownParamsMap = {};
-  KNOWN_PARAMS.forEach(value => knownParamsMap[value] = 1);
-  Object.keys(slot.params).forEach(key => {
-    if (!knownParamsMap[key]) {
-      ext[key] = slot.params[key];
-    }
-  });
-  return Object.keys(ext).length > 0 ? { prebid: ext } : null;
-}
 
 registerBidder(spec);
