+++ modules/pulsepointBidAdapter.js
@@ -1,6 +1,6 @@
 import { ortbConverter } from '../libraries/ortbConverter/converter.js';
-import {isArray} from '../src/utils.js';
-import {registerBidder} from '../src/adapters/bidderFactory.js';
+import { registerBidder } from '../src/adapters/bidderFactory.js';
+import { applyCommonImpParams } from '../libraries/impUtils.js';
 
 const DEFAULT_CURRENCY = 'USD';
 const KNOWN_PARAMS = ['cp', 'ct', 'cf', 'battr', 'deals'];
@@ -71,26 +71,7 @@ const converter = ortbConverter({
     const imp = buildImp(bidRequest, context);
     // tagid
     imp.tagid = bidRequest.params.ct.toString();
-    // unknown params
-    const unknownParams = slotUnknownParams(bidRequest);
-    if (imp.ext || unknownParams) {
-      imp.ext = Object.assign({}, imp.ext, unknownParams);
-    }
-    // battr
-    if (bidRequest.params.battr) {
-      ['banner', 'video', 'audio', 'native'].forEach(k => {
-        if (imp[k]) {
-          imp[k].battr = bidRequest.params.battr;
-        }
-      });
-    }
-    // deals
-    if (bidRequest.params.deals && isArray(bidRequest.params.deals)) {
-      imp.pmp = {
-        private_auction: 0,
-        deals: bidRequest.params.deals
-      };
-    }
+    applyCommonImpParams(imp, bidRequest, KNOWN_PARAMS);
     return imp;
   },
 
@@ -115,20 +96,4 @@ const converter = ortbConverter({
     return bidResponse;
   },
 });
-
-/**
- * Unknown params are captured and sent on ext
- */
-function slotUnknownParams(slot) {
-  const ext = {};
-  const knownParamsMap = {};
-  KNOWN_PARAMS.forEach(value => knownParamsMap[value] = 1);
-  Object.keys(slot.params).forEach(key => {
-    if (!knownParamsMap[key]) {
-      ext[key] = slot.params[key];
-    }
-  });
-  return Object.keys(ext).length > 0 ? { prebid: ext } : null;
-}
-
 registerBidder(spec);
